Reworked graphs
- Removed the templated create node method and NodeTypes
- Function to save all graphs is in utils
- Instead of using pointers as references use refs which are user defined.
  The create_vk_node method fixes the Vulkan handle issues and can be seen
  working from the framework graph (linking swapchain with vkImage and Image)
- Clean folder structure and add graph styling
