+++ framework/common/utils.cpp
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018-2019, Arm Limited and Contributors
+/* Copyright (c) 2018-2020, Arm Limited and Contributors
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -22,6 +22,8 @@
 
 #include "core/pipeline_layout.h"
 #include "core/shader_module.h"
+#include "graphing/framework_graph.h"
+#include "graphing/scene_graph.h"
 #include "scene_graph/components/image.h"
 #include "scene_graph/components/material.h"
 #include "scene_graph/components/mesh.h"
@@ -310,5 +312,24 @@ sg::Node &add_free_camera(sg::Scene &scene, const std::string &node_name, VkExte
 
 	return *camera_node;
 }
+
+namespace graphs
+{
+bool generate_all(RenderContext &context, sg::Scene &scene)
+{
+	bool success = true;
+	if (!graphing::framework_graph::generate(context))
+	{
+		LOGE("Failed to save render context graph");
+		success = false;
+	}
+	if (!graphing::scene_graph::generate(scene))
+	{
+		LOGE("Failed to save scene graph");
+		success = false;
+	}
+	return success;
+}
+}        // namespace graphs
 
 }        // namespace vkb
