+++ framework/common/vk_common.cpp
@@ -331,446 +331,6 @@ int32_t get_bits_per_pixel(VkFormat format)
 	}
 }
 
-const std::string to_string(VkFormat format)
-{
-	switch (format)
-	{
-		case VK_FORMAT_R4G4_UNORM_PACK8:
-			return "VK_FORMAT_R4G4_UNORM_PACK8";
-		case VK_FORMAT_R4G4B4A4_UNORM_PACK16:
-			return "VK_FORMAT_R4G4B4A4_UNORM_PACK16";
-		case VK_FORMAT_B4G4R4A4_UNORM_PACK16:
-			return "VK_FORMAT_B4G4R4A4_UNORM_PACK16";
-		case VK_FORMAT_R5G6B5_UNORM_PACK16:
-			return "VK_FORMAT_R5G6B5_UNORM_PACK16";
-		case VK_FORMAT_B5G6R5_UNORM_PACK16:
-			return "VK_FORMAT_B5G6R5_UNORM_PACK16";
-		case VK_FORMAT_R5G5B5A1_UNORM_PACK16:
-			return "VK_FORMAT_R5G5B5A1_UNORM_PACK16";
-		case VK_FORMAT_B5G5R5A1_UNORM_PACK16:
-			return "VK_FORMAT_B5G5R5A1_UNORM_PACK16";
-		case VK_FORMAT_A1R5G5B5_UNORM_PACK16:
-			return "VK_FORMAT_A1R5G5B5_UNORM_PACK16";
-		case VK_FORMAT_R8_UNORM:
-			return "VK_FORMAT_R8_UNORM";
-		case VK_FORMAT_R8_SNORM:
-			return "VK_FORMAT_R8_SNORM";
-		case VK_FORMAT_R8_USCALED:
-			return "VK_FORMAT_R8_USCALED";
-		case VK_FORMAT_R8_SSCALED:
-			return "VK_FORMAT_R8_SSCALED";
-		case VK_FORMAT_R8_UINT:
-			return "VK_FORMAT_R8_UINT";
-		case VK_FORMAT_R8_SINT:
-			return "VK_FORMAT_R8_SINT";
-		case VK_FORMAT_R8_SRGB:
-			return "VK_FORMAT_R8_SRGB";
-		case VK_FORMAT_R8G8_UNORM:
-			return "VK_FORMAT_R8G8_UNORM";
-		case VK_FORMAT_R8G8_SNORM:
-			return "VK_FORMAT_R8G8_SNORM";
-		case VK_FORMAT_R8G8_USCALED:
-			return "VK_FORMAT_R8G8_USCALED";
-		case VK_FORMAT_R8G8_SSCALED:
-			return "VK_FORMAT_R8G8_SSCALED";
-		case VK_FORMAT_R8G8_UINT:
-			return "VK_FORMAT_R8G8_UINT";
-		case VK_FORMAT_R8G8_SINT:
-			return "VK_FORMAT_R8G8_SINT";
-		case VK_FORMAT_R8G8_SRGB:
-			return "VK_FORMAT_R8G8_SRGB";
-		case VK_FORMAT_R8G8B8_UNORM:
-			return "VK_FORMAT_R8G8B8_UNORM";
-		case VK_FORMAT_R8G8B8_SNORM:
-			return "VK_FORMAT_R8G8B8_SNORM";
-		case VK_FORMAT_R8G8B8_USCALED:
-			return "VK_FORMAT_R8G8B8_USCALED";
-		case VK_FORMAT_R8G8B8_SSCALED:
-			return "VK_FORMAT_R8G8B8_SSCALED";
-		case VK_FORMAT_R8G8B8_UINT:
-			return "VK_FORMAT_R8G8B8_UINT";
-		case VK_FORMAT_R8G8B8_SINT:
-			return "VK_FORMAT_R8G8B8_SINT";
-		case VK_FORMAT_R8G8B8_SRGB:
-			return "VK_FORMAT_R8G8B8_SRGB";
-		case VK_FORMAT_B8G8R8_UNORM:
-			return "VK_FORMAT_B8G8R8_UNORM";
-		case VK_FORMAT_B8G8R8_SNORM:
-			return "VK_FORMAT_B8G8R8_SNORM";
-		case VK_FORMAT_B8G8R8_USCALED:
-			return "VK_FORMAT_B8G8R8_USCALED";
-		case VK_FORMAT_B8G8R8_SSCALED:
-			return "VK_FORMAT_B8G8R8_SSCALED";
-		case VK_FORMAT_B8G8R8_UINT:
-			return "VK_FORMAT_B8G8R8_UINT";
-		case VK_FORMAT_B8G8R8_SINT:
-			return "VK_FORMAT_B8G8R8_SINT";
-		case VK_FORMAT_B8G8R8_SRGB:
-			return "VK_FORMAT_B8G8R8_SRGB";
-		case VK_FORMAT_R8G8B8A8_UNORM:
-			return "VK_FORMAT_R8G8B8A8_UNORM";
-		case VK_FORMAT_R8G8B8A8_SNORM:
-			return "VK_FORMAT_R8G8B8A8_SNORM";
-		case VK_FORMAT_R8G8B8A8_USCALED:
-			return "VK_FORMAT_R8G8B8A8_USCALED";
-		case VK_FORMAT_R8G8B8A8_SSCALED:
-			return "VK_FORMAT_R8G8B8A8_SSCALED";
-		case VK_FORMAT_R8G8B8A8_UINT:
-			return "VK_FORMAT_R8G8B8A8_UINT";
-		case VK_FORMAT_R8G8B8A8_SINT:
-			return "VK_FORMAT_R8G8B8A8_SINT";
-		case VK_FORMAT_R8G8B8A8_SRGB:
-			return "VK_FORMAT_R8G8B8A8_SRGB";
-		case VK_FORMAT_B8G8R8A8_UNORM:
-			return "VK_FORMAT_B8G8R8A8_UNORM";
-		case VK_FORMAT_B8G8R8A8_SNORM:
-			return "VK_FORMAT_B8G8R8A8_SNORM";
-		case VK_FORMAT_B8G8R8A8_USCALED:
-			return "VK_FORMAT_B8G8R8A8_USCALED";
-		case VK_FORMAT_B8G8R8A8_SSCALED:
-			return "VK_FORMAT_B8G8R8A8_SSCALED";
-		case VK_FORMAT_B8G8R8A8_UINT:
-			return "VK_FORMAT_B8G8R8A8_UINT";
-		case VK_FORMAT_B8G8R8A8_SINT:
-			return "VK_FORMAT_B8G8R8A8_SINT";
-		case VK_FORMAT_B8G8R8A8_SRGB:
-			return "VK_FORMAT_B8G8R8A8_SRGB";
-		case VK_FORMAT_A8B8G8R8_UNORM_PACK32:
-			return "VK_FORMAT_A8B8G8R8_UNORM_PACK32";
-		case VK_FORMAT_A8B8G8R8_SNORM_PACK32:
-			return "VK_FORMAT_A8B8G8R8_SNORM_PACK32";
-		case VK_FORMAT_A8B8G8R8_USCALED_PACK32:
-			return "VK_FORMAT_A8B8G8R8_USCALED_PACK32";
-		case VK_FORMAT_A8B8G8R8_SSCALED_PACK32:
-			return "VK_FORMAT_A8B8G8R8_SSCALED_PACK32";
-		case VK_FORMAT_A8B8G8R8_UINT_PACK32:
-			return "VK_FORMAT_A8B8G8R8_UINT_PACK32";
-		case VK_FORMAT_A8B8G8R8_SINT_PACK32:
-			return "VK_FORMAT_A8B8G8R8_SINT_PACK32";
-		case VK_FORMAT_A8B8G8R8_SRGB_PACK32:
-			return "VK_FORMAT_A8B8G8R8_SRGB_PACK32";
-		case VK_FORMAT_A2R10G10B10_UNORM_PACK32:
-			return "VK_FORMAT_A2R10G10B10_UNORM_PACK32";
-		case VK_FORMAT_A2R10G10B10_SNORM_PACK32:
-			return "VK_FORMAT_A2R10G10B10_SNORM_PACK32";
-		case VK_FORMAT_A2R10G10B10_USCALED_PACK32:
-			return "VK_FORMAT_A2R10G10B10_USCALED_PACK32";
-		case VK_FORMAT_A2R10G10B10_SSCALED_PACK32:
-			return "VK_FORMAT_A2R10G10B10_SSCALED_PACK32";
-		case VK_FORMAT_A2R10G10B10_UINT_PACK32:
-			return "VK_FORMAT_A2R10G10B10_UINT_PACK32";
-		case VK_FORMAT_A2R10G10B10_SINT_PACK32:
-			return "VK_FORMAT_A2R10G10B10_SINT_PACK32";
-		case VK_FORMAT_A2B10G10R10_UNORM_PACK32:
-			return "VK_FORMAT_A2B10G10R10_UNORM_PACK32";
-		case VK_FORMAT_A2B10G10R10_SNORM_PACK32:
-			return "VK_FORMAT_A2B10G10R10_SNORM_PACK32";
-		case VK_FORMAT_A2B10G10R10_USCALED_PACK32:
-			return "VK_FORMAT_A2B10G10R10_USCALED_PACK32";
-		case VK_FORMAT_A2B10G10R10_SSCALED_PACK32:
-			return "VK_FORMAT_A2B10G10R10_SSCALED_PACK32";
-		case VK_FORMAT_A2B10G10R10_UINT_PACK32:
-			return "VK_FORMAT_A2B10G10R10_UINT_PACK32";
-		case VK_FORMAT_A2B10G10R10_SINT_PACK32:
-			return "VK_FORMAT_A2B10G10R10_SINT_PACK32";
-		case VK_FORMAT_R16_UNORM:
-			return "VK_FORMAT_R16_UNORM";
-		case VK_FORMAT_R16_SNORM:
-			return "VK_FORMAT_R16_SNORM";
-		case VK_FORMAT_R16_USCALED:
-			return "VK_FORMAT_R16_USCALED";
-		case VK_FORMAT_R16_SSCALED:
-			return "VK_FORMAT_R16_SSCALED";
-		case VK_FORMAT_R16_UINT:
-			return "VK_FORMAT_R16_UINT";
-		case VK_FORMAT_R16_SINT:
-			return "VK_FORMAT_R16_SINT";
-		case VK_FORMAT_R16_SFLOAT:
-			return "VK_FORMAT_R16_SFLOAT";
-		case VK_FORMAT_R16G16_UNORM:
-			return "VK_FORMAT_R16G16_UNORM";
-		case VK_FORMAT_R16G16_SNORM:
-			return "VK_FORMAT_R16G16_SNORM";
-		case VK_FORMAT_R16G16_USCALED:
-			return "VK_FORMAT_R16G16_USCALED";
-		case VK_FORMAT_R16G16_SSCALED:
-			return "VK_FORMAT_R16G16_SSCALED";
-		case VK_FORMAT_R16G16_UINT:
-			return "VK_FORMAT_R16G16_UINT";
-		case VK_FORMAT_R16G16_SINT:
-			return "VK_FORMAT_R16G16_SINT";
-		case VK_FORMAT_R16G16_SFLOAT:
-			return "VK_FORMAT_R16G16_SFLOAT";
-		case VK_FORMAT_R16G16B16_UNORM:
-			return "VK_FORMAT_R16G16B16_UNORM";
-		case VK_FORMAT_R16G16B16_SNORM:
-			return "VK_FORMAT_R16G16B16_SNORM";
-		case VK_FORMAT_R16G16B16_USCALED:
-			return "VK_FORMAT_R16G16B16_USCALED";
-		case VK_FORMAT_R16G16B16_SSCALED:
-			return "VK_FORMAT_R16G16B16_SSCALED";
-		case VK_FORMAT_R16G16B16_UINT:
-			return "VK_FORMAT_R16G16B16_UINT";
-		case VK_FORMAT_R16G16B16_SINT:
-			return "VK_FORMAT_R16G16B16_SINT";
-		case VK_FORMAT_R16G16B16_SFLOAT:
-			return "VK_FORMAT_R16G16B16_SFLOAT";
-		case VK_FORMAT_R16G16B16A16_UNORM:
-			return "VK_FORMAT_R16G16B16A16_UNORM";
-		case VK_FORMAT_R16G16B16A16_SNORM:
-			return "VK_FORMAT_R16G16B16A16_SNORM";
-		case VK_FORMAT_R16G16B16A16_USCALED:
-			return "VK_FORMAT_R16G16B16A16_USCALED";
-		case VK_FORMAT_R16G16B16A16_SSCALED:
-			return "VK_FORMAT_R16G16B16A16_SSCALED";
-		case VK_FORMAT_R16G16B16A16_UINT:
-			return "VK_FORMAT_R16G16B16A16_UINT";
-		case VK_FORMAT_R16G16B16A16_SINT:
-			return "VK_FORMAT_R16G16B16A16_SINT";
-		case VK_FORMAT_R16G16B16A16_SFLOAT:
-			return "VK_FORMAT_R16G16B16A16_SFLOAT";
-		case VK_FORMAT_R32_UINT:
-			return "VK_FORMAT_R32_UINT";
-		case VK_FORMAT_R32_SINT:
-			return "VK_FORMAT_R32_SINT";
-		case VK_FORMAT_R32_SFLOAT:
-			return "VK_FORMAT_R32_SFLOAT";
-		case VK_FORMAT_R32G32_UINT:
-			return "VK_FORMAT_R32G32_UINT";
-		case VK_FORMAT_R32G32_SINT:
-			return "VK_FORMAT_R32G32_SINT";
-		case VK_FORMAT_R32G32_SFLOAT:
-			return "VK_FORMAT_R32G32_SFLOAT";
-		case VK_FORMAT_R32G32B32_UINT:
-			return "VK_FORMAT_R32G32B32_UINT";
-		case VK_FORMAT_R32G32B32_SINT:
-			return "VK_FORMAT_R32G32B32_SINT";
-		case VK_FORMAT_R32G32B32_SFLOAT:
-			return "VK_FORMAT_R32G32B32_SFLOAT";
-		case VK_FORMAT_R32G32B32A32_UINT:
-			return "VK_FORMAT_R32G32B32A32_UINT";
-		case VK_FORMAT_R32G32B32A32_SINT:
-			return "VK_FORMAT_R32G32B32A32_SINT";
-		case VK_FORMAT_R32G32B32A32_SFLOAT:
-			return "VK_FORMAT_R32G32B32A32_SFLOAT";
-		case VK_FORMAT_R64_UINT:
-			return "VK_FORMAT_R64_UINT";
-		case VK_FORMAT_R64_SINT:
-			return "VK_FORMAT_R64_SINT";
-		case VK_FORMAT_R64_SFLOAT:
-			return "VK_FORMAT_R64_SFLOAT";
-		case VK_FORMAT_R64G64_UINT:
-			return "VK_FORMAT_R64G64_UINT";
-		case VK_FORMAT_R64G64_SINT:
-			return "VK_FORMAT_R64G64_SINT";
-		case VK_FORMAT_R64G64_SFLOAT:
-			return "VK_FORMAT_R64G64_SFLOAT";
-		case VK_FORMAT_R64G64B64_UINT:
-			return "VK_FORMAT_R64G64B64_UINT";
-		case VK_FORMAT_R64G64B64_SINT:
-			return "VK_FORMAT_R64G64B64_SINT";
-		case VK_FORMAT_R64G64B64_SFLOAT:
-			return "VK_FORMAT_R64G64B64_SFLOAT";
-		case VK_FORMAT_R64G64B64A64_UINT:
-			return "VK_FORMAT_R64G64B64A64_UINT";
-		case VK_FORMAT_R64G64B64A64_SINT:
-			return "VK_FORMAT_R64G64B64A64_SINT";
-		case VK_FORMAT_R64G64B64A64_SFLOAT:
-			return "VK_FORMAT_R64G64B64A64_SFLOAT";
-		case VK_FORMAT_B10G11R11_UFLOAT_PACK32:
-			return "VK_FORMAT_B10G11R11_UFLOAT_PACK32";
-		case VK_FORMAT_E5B9G9R9_UFLOAT_PACK32:
-			return "VK_FORMAT_E5B9G9R9_UFLOAT_PACK32";
-		case VK_FORMAT_D16_UNORM:
-			return "VK_FORMAT_D16_UNORM";
-		case VK_FORMAT_X8_D24_UNORM_PACK32:
-			return "VK_FORMAT_X8_D24_UNORM_PACK32";
-		case VK_FORMAT_D32_SFLOAT:
-			return "VK_FORMAT_D32_SFLOAT";
-		case VK_FORMAT_S8_UINT:
-			return "VK_FORMAT_S8_UINT";
-		case VK_FORMAT_D16_UNORM_S8_UINT:
-			return "VK_FORMAT_D16_UNORM_S8_UINT";
-		case VK_FORMAT_D24_UNORM_S8_UINT:
-			return "VK_FORMAT_D24_UNORM_S8_UINT";
-		case VK_FORMAT_D32_SFLOAT_S8_UINT:
-			return "VK_FORMAT_D32_SFLOAT_S8_UINT";
-		case VK_FORMAT_UNDEFINED:
-			return "VK_FORMAT_UNDEFINED";
-		default:
-			return "VK_FORMAT_INVALID";
-	}
-}
-
-const std::string to_string(VkPresentModeKHR present_mode)
-{
-	switch (present_mode)
-	{
-		case VK_PRESENT_MODE_MAILBOX_KHR:
-			return "VK_PRESENT_MODE_MAILBOX_KHR";
-		case VK_PRESENT_MODE_IMMEDIATE_KHR:
-			return "VK_PRESENT_MODE_IMMEDIATE_KHR";
-		case VK_PRESENT_MODE_FIFO_KHR:
-			return "VK_PRESENT_MODE_FIFO_KHR";
-		case VK_PRESENT_MODE_FIFO_RELAXED_KHR:
-			return "VK_PRESENT_MODE_FIFO_RELAXED_KHR";
-		case VK_PRESENT_MODE_SHARED_CONTINUOUS_REFRESH_KHR:
-			return "VK_PRESENT_MODE_SHARED_CONTINUOUS_REFRESH_KHR";
-		case VK_PRESENT_MODE_SHARED_DEMAND_REFRESH_KHR:
-			return "VK_PRESENT_MODE_SHARED_DEMAND_REFRESH_KHR";
-		default:
-			return "Unknown Present Mode";
-	}
-}
-
-const std::string to_string(VkResult result)
-{
-	switch (result)
-	{
-#define STR(r)   \
-	case VK_##r: \
-		return #r
-		STR(NOT_READY);
-		STR(TIMEOUT);
-		STR(EVENT_SET);
-		STR(EVENT_RESET);
-		STR(INCOMPLETE);
-		STR(ERROR_OUT_OF_HOST_MEMORY);
-		STR(ERROR_OUT_OF_DEVICE_MEMORY);
-		STR(ERROR_INITIALIZATION_FAILED);
-		STR(ERROR_DEVICE_LOST);
-		STR(ERROR_MEMORY_MAP_FAILED);
-		STR(ERROR_LAYER_NOT_PRESENT);
-		STR(ERROR_EXTENSION_NOT_PRESENT);
-		STR(ERROR_FEATURE_NOT_PRESENT);
-		STR(ERROR_INCOMPATIBLE_DRIVER);
-		STR(ERROR_TOO_MANY_OBJECTS);
-		STR(ERROR_FORMAT_NOT_SUPPORTED);
-		STR(ERROR_SURFACE_LOST_KHR);
-		STR(ERROR_NATIVE_WINDOW_IN_USE_KHR);
-		STR(SUBOPTIMAL_KHR);
-		STR(ERROR_OUT_OF_DATE_KHR);
-		STR(ERROR_INCOMPATIBLE_DISPLAY_KHR);
-		STR(ERROR_VALIDATION_FAILED_EXT);
-		STR(ERROR_INVALID_SHADER_NV);
-#undef STR
-		default:
-			return "UNKNOWN_ERROR";
-	}
-}
-
-const std::string to_string(VkPhysicalDeviceType type)
-{
-	switch (type)
-	{
-#define STR(r)                        \
-	case VK_PHYSICAL_DEVICE_TYPE_##r: \
-		return #r
-		STR(OTHER);
-		STR(INTEGRATED_GPU);
-		STR(DISCRETE_GPU);
-		STR(VIRTUAL_GPU);
-#undef STR
-		default:
-			return "UNKNOWN_DEVICE_TYPE";
-	}
-}
-
-const std::string to_string(VkSurfaceTransformFlagBitsKHR transform_flag)
-{
-	switch (transform_flag)
-	{
-		case VK_SURFACE_TRANSFORM_IDENTITY_BIT_KHR:
-			return "SURFACE_TRANSFORM_IDENTITY";
-		case VK_SURFACE_TRANSFORM_ROTATE_90_BIT_KHR:
-			return "SURFACE_TRANSFORM_ROTATE_90";
-		case VK_SURFACE_TRANSFORM_ROTATE_180_BIT_KHR:
-			return "SURFACE_TRANSFORM_ROTATE_180";
-		case VK_SURFACE_TRANSFORM_ROTATE_270_BIT_KHR:
-			return "SURFACE_TRANSFORM_ROTATE_270";
-		case VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_BIT_KHR:
-			return "SURFACE_TRANSFORM_HORIZONTAL_MIRROR";
-		case VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90_BIT_KHR:
-			return "SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_90";
-		case VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180_BIT_KHR:
-			return "SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_180";
-		case VK_SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270_BIT_KHR:
-			return "SURFACE_TRANSFORM_HORIZONTAL_MIRROR_ROTATE_270";
-		case VK_SURFACE_TRANSFORM_INHERIT_BIT_KHR:
-			return "SURFACE_TRANSFORM_INHERIT";
-		case VK_SURFACE_TRANSFORM_FLAG_BITS_MAX_ENUM_KHR:
-			return "SURFACE_TRANSFORM_FLAG_BITS_MAX_ENUM";
-		default:
-			return "[Unknown transform flag]";
-	}
-}
-
-const std::string to_string(VkSurfaceFormatKHR surface_format)
-{
-	std::string surface_format_string = to_string(surface_format.format) + ", ";
-
-	switch (surface_format.colorSpace)
-	{
-		case VK_COLORSPACE_SRGB_NONLINEAR_KHR:
-			surface_format_string += "VK_COLORSPACE_SRGB_NONLINEAR_KHR";
-			break;
-		default:
-			surface_format_string += "UNKNOWN COLOR SPACE";
-	}
-	return surface_format_string;
-}
-
-const std::string to_string(VkCompositeAlphaFlagBitsKHR composite_alpha)
-{
-	switch (composite_alpha)
-	{
-		case VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR:
-			return "VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR";
-		case VK_COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR:
-			return "VK_COMPOSITE_ALPHA_PRE_MULTIPLIED_BIT_KHR";
-		case VK_COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR:
-			return "VK_COMPOSITE_ALPHA_POST_MULTIPLIED_BIT_KHR";
-		case VK_COMPOSITE_ALPHA_INHERIT_BIT_KHR:
-			return "VK_COMPOSITE_ALPHA_INHERIT_BIT_KHR";
-		case VK_COMPOSITE_ALPHA_FLAG_BITS_MAX_ENUM_KHR:
-			return "VK_COMPOSITE_ALPHA_FLAG_BITS_MAX_ENUM_KHR";
-		default:
-			return "UNKNOWN COMPOSITE ALPHA FLAG";
-	}
-}
-
-const std::string to_string(VkImageUsageFlagBits image_usage)
-{
-	switch (image_usage)
-	{
-		case VK_IMAGE_USAGE_TRANSFER_SRC_BIT:
-			return "VK_IMAGE_USAGE_TRANSFER_SRC_BIT";
-		case VK_IMAGE_USAGE_TRANSFER_DST_BIT:
-			return "VK_IMAGE_USAGE_TRANSFER_DST_BIT";
-		case VK_IMAGE_USAGE_SAMPLED_BIT:
-			return "VK_IMAGE_USAGE_SAMPLED_BIT";
-		case VK_IMAGE_USAGE_STORAGE_BIT:
-			return "VK_IMAGE_USAGE_STORAGE_BIT";
-		case VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT:
-			return "VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT";
-		case VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT:
-			return "VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT";
-		case VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT:
-			return "VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT";
-		case VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT:
-			return "VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT";
-		case VK_IMAGE_USAGE_FLAG_BITS_MAX_ENUM:
-			return "VK_IMAGE_USAGE_FLAG_BITS_MAX_ENUM";
-		default:
-			return "UNKNOWN IMAGE USAGE FLAG";
-	}
-}
-
-const std::string to_string(VkExtent2D extent)
-{
-	return fmt::format("{}x{}", extent.width, extent.height);
-}
-
 VkShaderModule load_shader(const std::string &filename, VkDevice device, VkShaderStageFlagBits stage)
 {
 	vkb::GLSLCompiler glsl_compiler;
