+++ framework/common/vk_common.h
@@ -89,69 +89,6 @@ bool is_buffer_descriptor_type(VkDescriptorType descriptor_type);
 int32_t get_bits_per_pixel(VkFormat format);
 
 /**
- * @brief Helper function to convert a VkFormat enum to a string
- * @param format Vulkan format to convert.
- * @return The string to return.
- */
-const std::string to_string(VkFormat format);
-
-/**
- * @brief Helper function to convert a VkPresentModeKHR to a string
- * @param present_mode Vulkan present mode to convert.
- * @return The string to return.
- */
-const std::string to_string(VkPresentModeKHR present_mode);
-
-/**
- * @brief Helper function to convert a VkResult enum to a string
- * @param format Vulkan result to convert.
- * @return The string to return.
- */
-const std::string to_string(VkResult result);
-
-/**
- * @brief Helper function to convert a VkPhysicalDeviceType enum to a string
- * @param format Vulkan physical device type to convert.
- * @return The string to return.
- */
-const std::string to_string(VkPhysicalDeviceType type);
-
-/**
- * @brief Helper function to convert a VkSurfaceTransformFlagBitsKHR flag to a string
- * @param transform_flag Vulkan surface transform flag bit to convert.
- * @return The string to return.
- */
-const std::string to_string(VkSurfaceTransformFlagBitsKHR transform_flag);
-
-/**
- * @brief Helper function to convert a VkSurfaceFormatKHR format to a string
- * @param surface_format Vulkan surface format to convert.
- * @return The string to return.
- */
-const std::string to_string(VkSurfaceFormatKHR surface_format);
-
-/**
- * @brief Helper function to convert a VkCompositeAlphaFlagBitsKHR flag to a string
- * @param composite_alpha Vulkan composite alpha flag bit to convert.
- * @return The string to return.
- */
-const std::string to_string(VkCompositeAlphaFlagBitsKHR composite_alpha);
-
-/**
- * @brief Helper function to convert a VkImageUsageFlagBits flag to a string
- * @param image_usage Vulkan image usage flag bit to convert.
- * @return The string to return.
- */
-const std::string to_string(VkImageUsageFlagBits image_usage);
-
-/**
- * @brief Helper function to convert a VkExtent2D flag to a string
- * @param extent Vulkan extent to convert.
- * @return The string to return.
- */
-const std::string to_string(VkExtent2D extent);
-
-/**
  * @brief Helper function to create a VkShaderModule
  * @param filename The shader location
  * @param device The logical device
