+++ framework/gui.cpp
@@ -43,7 +43,6 @@ VKBP_ENABLE_WARNINGS()
 #include "platform/window.h"
 #include "rendering/render_context.h"
 #include "timer.h"
-#include "utils/graphs.h"
 #include "vulkan_sample.h"
 
 namespace vkb
@@ -849,7 +848,7 @@ void Gui::show_debug_window(DebugInfo &debug_info, const ImVec2 &position)
 
 	if (ImGui::Button("Save Debug Graphs"))
 	{
-		if (utils::debug_graphs(sample.get_render_context(), sample.get_scene()))
+		if (graphs::generate_all(sample.get_render_context(), sample.get_scene()))
 		{
 			message = "Graphs Saved!";
 		}
